feat: override the configurations of the http client
Override the configurations of the http client connection to increase the timeout. This is a hacky solution to give more time to the proxy server to finish the swapping operations and reply. This is to avoid introducing a big refactor to submit tasks to the server and listen to the status before getting the response back via another endpoint.

In order for this to take effect, the service needs to be bumped to the pending autonomy release in https://github.com/valory-xyz/open-autonomy/pull/2216, which utilizes `open-aea v1.51.0`, which adds the support for the timeout overriding to the connection in https://github.com/valory-xyz/open-aea/pull/730.